libgep: access GepChannel vector iff lock is held
We want to protect GepChannelArray::SendMessage() callers to ensure
the contents of the vector are not changed while they are looping
through it. This requires adding a mutex lock in these functions.

We are also changing the vector access mutex to recursive. The
previous (implicit) approach was to lock the mutex on receive, but
not on transmit: The rationale was that some users may not realize
that their Recv<Message>() function is run in a callback run by
the GepChannelArray::RecvData(), which does hold the vector lock.
If they try to send messages, they will try to get the lock again,
and lock. For example, one of our tests uses the receive callback
to send another message (ping/pong protocol).

Google-Bug-Id: b/26245616

Change-Id: Ia48774af60fe08a44439ef745cf0aefea17b7cc4