libgep: export client ID to server
We want to export the client ID to the server, so the latter can
know which client sent a message. This includes adding a per-client
unique ID to all the receive calls, and AddClient/DelClient callbacks
that the server can use to learn when there is a new client or an
old client has left.

Also made Gep[Client|Server]::Start/Stop functions virtual.

Change-Id: I3e5b7c0c1cb1726c45c4c3c9c256d8c42cdd09fc